pyproject.toml
[tool.poetry]
name = "project-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
black = "^24.4.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
